Split CSS into a chunk of its own
Hashing seems to work as expected now. Earlier altering CSS changed the
app chunk as well -> both js/css got invalidated. The only problem with
this approach is that it generates an empty JS file for the entry and
there doesn't seem to be an easy way to avoid that.